refactor: add useDataDestinationsWithReports hook to fetch data destinations and reports concurrently
