+++ apps/web/src/features/data-destination/shared/model/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './useDataDestination';
 export * from './useDataDestinationVisibility';
+export * from './useDataDestinationsWithReports';
