+++ apps/web/src/pages/data-marts/edit/DataMartDestinationsContent.tsx
@@ -1,5 +1,3 @@
-import { useEffect, useState } from 'react';
-import { useReport } from '../../../features/data-marts/reports/shared/model/hooks';
 import {
   DestinationCard,
   EmptyDataMartDestinationsState,
@@ -7,51 +5,27 @@ import {
 import { useOutletContext } from 'react-router-dom';
 import type { DataMartContextType } from '../../../features/data-marts/edit/model/context/types';
 import { SkeletonList } from '@owox/ui/components/common/skeleton-list';
-import { useDataDestination } from '../../../features/data-destination/shared';
-import { DataDestinationProvider } from '../../../features/data-destination/shared';
+import {
+  useDataDestinationsWithReports,
+  DataDestinationProvider,
+  type DataDestination,
+} from '../../../features/data-destination/shared';
 import { ReportsProvider } from '../../../features/data-marts/reports/shared/model/context';
 
 function DataMartDestinationsContentInner() {
   const { dataMart } = useOutletContext<DataMartContextType>();
-  const { fetchReportsByDataMartId } = useReport();
-  const {
-    dataDestinations,
-    fetchDataDestinations,
-    loading: destinationsLoading,
-  } = useDataDestination();
+  const { dataDestinations, isLoading } = useDataDestinationsWithReports();
 
-  // Add loading state management
-  const [isLoading, setIsLoading] = useState(true);
-
-  // Fetch destinations and reports when component mounts
-  useEffect(() => {
-    const fetchData = async () => {
-      setIsLoading(true);
-      try {
-        if (!dataMart) return;
-        await Promise.all([fetchDataDestinations(), fetchReportsByDataMartId(dataMart.id)]);
-      } catch (err) {
-        console.error(err);
-      } finally {
-        setIsLoading(false);
-      }
-    };
-
-    void fetchData();
-  }, [fetchDataDestinations, fetchReportsByDataMartId, dataMart]);
-
-  if (!dataMart) {
-    return null;
-  }
+  if (!dataMart) return null;
 
   return (
     <div className='flex flex-col gap-4'>
-      {isLoading || destinationsLoading ? (
+      {isLoading ? (
         <SkeletonList />
       ) : dataDestinations.length === 0 ? (
         <EmptyDataMartDestinationsState />
       ) : (
-        dataDestinations.map(destination => (
+        dataDestinations.map((destination: DataDestination) => (
           <DestinationCard
             key={destination.id}
             destination={destination}
